+++ engine/src/com/twinsoft/convertigo/engine/util/XMLUtils.java
@@ -46,6 +46,7 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.parsers.SAXParser;
 import javax.xml.parsers.SAXParserFactory;
+import javax.xml.XMLConstants;
 import javax.xml.transform.OutputKeys;
 import javax.xml.transform.Result;
 import javax.xml.transform.Source;
@@ -89,6 +90,19 @@ public class XMLUtils {
 		@Override
 		protected DocumentBuilderFactory initialValue() {
 			DocumentBuilderFactory documentBuilderFactory = DocumentBuilderFactory.newInstance();
+			// Disable external entity resolution and external DTD access. Internal
+			// DTD subsets (used by some shipped XSL) keep working; only external
+			// references are blocked.
+			try {
+				documentBuilderFactory.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
+				documentBuilderFactory.setFeature("http://xml.org/sax/features/external-general-entities", false);
+				documentBuilderFactory.setFeature("http://xml.org/sax/features/external-parameter-entities", false);
+				documentBuilderFactory.setFeature("http://apache.org/xml/features/nonvalidating/load-external-dtd", false);
+				documentBuilderFactory.setXIncludeAware(false);
+				documentBuilderFactory.setExpandEntityReferences(false);
+			} catch (Exception e) {
+				Engine.logEngine.warn("Unable to harden the XML document builder factory: " + e.getMessage());
+			}
 			try {
 				String s = EnginePropertiesManager.getProperty(PropertyName.DOCUMENT_NAMESPACE_AWARE);
 				if (s.equalsIgnoreCase("true"))
@@ -947,8 +961,29 @@ public class XMLUtils {
 		return parseDOM(new File(filename));
 	}
 
+	// Builder used to parse XML coming from untrusted string input (e.g. request
+	// parameters). It forbids DOCTYPE declarations entirely, so no entity can be
+	// defined or referenced.
+	private static ThreadLocal<DocumentBuilder> secureDocumentBuilder = new ThreadLocal<DocumentBuilder>() {
+		@Override
+		protected DocumentBuilder initialValue() {
+			try {
+				DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
+				factory.setFeature("http://apache.org/xml/features/disallow-doctype-decl", true);
+				factory.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
+				factory.setXIncludeAware(false);
+				factory.setExpandEntityReferences(false);
+				factory.setNamespaceAware(true);
+				return factory.newDocumentBuilder();
+			} catch (ParserConfigurationException e) {
+				Engine.logEngine.error("Unable to create the secure XML document builder", e);
+				return null;
+			}
+		}
+	};
+
 	static public Document parseDOMFromString(String sDom) throws SAXException, IOException {
-		Document dom = getDefaultDocumentBuilder().parse(new InputSource(new StringReader(sDom)));
+		Document dom = secureDocumentBuilder.get().parse(new InputSource(new StringReader(sDom)));
 		return dom;
 	}